requirements.txt
asgiref==3.3.1
boto3==1.16.33
botocore==1.19.33
certifi==2020.12.5
cffi==1.14.4
chardet==4.0.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.3.1
defusedxml==0.6.0
Django==3.0
django-allauth==0.44.0
django-auth==0.1.9
django-debug-toolbar==3.2
django-rest-auth==0.9.5
django-storages==1.11.1
djangorestframework==3.10.3
djangorestframework-jwt==1.11.0
drf-yasg==1.20.0
idna==2.10
inflection==0.5.1
install==1.3.4
itypes==1.2.0
Jinja2==2.11.3
jmespath==0.10.0
MarkupSafe==1.1.1
mysqlclient==2.0.3
oauthlib==3.1.0
packaging==20.9
Pillow==8.0.1
pycparser==2.20
PyJWT==1.7.1
PyMySQL==1.0.2
pyparsing==2.4.7
python-dateutil==2.8.1
python3-openid==3.2.0
pytz==2020.4
requests==2.25.1
requests-oauthlib==1.3.0
ruamel.yaml==0.16.12
ruamel.yaml.clib==0.2.2
s3transfer==0.3.3
six==1.15.0
sqlparse==0.4.1
uritemplate==3.0.1
urllib3==1.26.2
celery==4.4